requirements.txt
click==7.1.2
fix==1.0.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
nltk==3.3
protobuf==3.13.0
six==1.15.0
SQLAlchemy==1.3.20
touch==0.0.0
values==0.0.0
Werkzeug==1.0.1
